# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Lindéro API",
      "description": "The Lindéro project allows you to sell and buy paintings, sculptures and photographs. Sales are carried out through auction or at a fixed price.",
      "contact": {
        "name": "Elena Sheina",
        "email": "lena_sheina@bk.ru"
      },
      "version": "0.0.1"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Dev server"
      }
    ],
    "paths": {
...
      "/card": {
        "put": {
          "tags": [
            "Банковская карта"
          ],
          "summary": "Изменить существующую карту",
          "operationId": "changeCard",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "403": {
              "description": "Действие запрещено",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ForbiddenActionException"
                  }
                }
              }
            },
            "200": {
              "description": "Адрес успешно изменен",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Банковская карта"
          ],
          "summary": "Добавить новую карту",
          "operationId": "addCard",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardNewDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Карта успешно добавлена",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Банковская карта"
          ],
          "summary": "Удалить существующую карту",
          "operationId": "deleteCard",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntIdRequestDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Карта успешно удалена",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "403": {
              "description": "Действие запрещено",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ForbiddenActionException"
                  }
                }
              }
            }
          }
        }
      },
...
      "/cards": {
        "get": {
          "tags": [
            "Банковская карта"
          ],
          "summary": "Получить все сохраненные карты пользователя",
          "operationId": "getCards",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список карт",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CardDTO"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "BadCredentialsException": {
          "type": "object",
          "properties": {
            "cause": {
              "type": "object",
              "properties": {
                "stackTrace": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "classLoaderName": {
                        "type": "string"
                      },
                      "moduleName": {
                        "type": "string"
                      },
                      "moduleVersion": {
                        "type": "string"
                      },
                      "methodName": {
                        "type": "string"
                      },
                      "fileName": {
                        "type": "string"
                      },
                      "lineNumber": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "nativeMethod": {
                        "type": "boolean"
                      },
                      "className": {
                        "type": "string"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "suppressed": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "stackTrace": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "classLoaderName": {
                              "type": "string"
                            },
                            "moduleName": {
                              "type": "string"
                            },
                            "moduleVersion": {
                              "type": "string"
                            },
                            "methodName": {
                              "type": "string"
                            },
                            "fileName": {
                              "type": "string"
                            },
                            "lineNumber": {
                              "type": "integer",
                              "format": "int32"
                            },
                            "nativeMethod": {
                              "type": "boolean"
                            },
                            "className": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "message": {
                        "type": "string"
                      },
                      "localizedMessage": {
                        "type": "string"
                      }
                    }
                  }
                },
                "localizedMessage": {
                  "type": "string"
                }
              }
            },
            "stackTrace": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "classLoaderName": {
                    "type": "string"
                  },
                  "moduleName": {
                    "type": "string"
                  },
                  "moduleVersion": {
                    "type": "string"
                  },
                  "methodName": {
                    "type": "string"
                  },
                  "fileName": {
                    "type": "string"
                  },
                  "lineNumber": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "nativeMethod": {
                    "type": "boolean"
                  },
                  "className": {
                    "type": "string"
                  }
                }
              }
            },
            "status": {
              "type": "string",
              "enum": [
                "100 CONTINUE",
                "101 SWITCHING_PROTOCOLS",
                "102 PROCESSING",
                "103 EARLY_HINTS",
                "103 CHECKPOINT",
                "200 OK",
                "201 CREATED",
                "202 ACCEPTED",
                "203 NON_AUTHORITATIVE_INFORMATION",
                "204 NO_CONTENT",
                "205 RESET_CONTENT",
                "206 PARTIAL_CONTENT",
                "207 MULTI_STATUS",
                "208 ALREADY_REPORTED",
                "226 IM_USED",
                "300 MULTIPLE_CHOICES",
                "301 MOVED_PERMANENTLY",
                "302 FOUND",
                "302 MOVED_TEMPORARILY",
                "303 SEE_OTHER",
                "304 NOT_MODIFIED",
                "305 USE_PROXY",
                "307 TEMPORARY_REDIRECT",
                "308 PERMANENT_REDIRECT",
                "400 BAD_REQUEST",
                "401 UNAUTHORIZED",
                "402 PAYMENT_REQUIRED",
                "403 FORBIDDEN",
                "404 NOT_FOUND",
                "405 METHOD_NOT_ALLOWED",
                "406 NOT_ACCEPTABLE",
                "407 PROXY_AUTHENTICATION_REQUIRED",
                "408 REQUEST_TIMEOUT",
                "409 CONFLICT",
                "410 GONE",
                "411 LENGTH_REQUIRED",
                "412 PRECONDITION_FAILED",
                "413 PAYLOAD_TOO_LARGE",
                "413 REQUEST_ENTITY_TOO_LARGE",
                "414 URI_TOO_LONG",
                "414 REQUEST_URI_TOO_LONG",
                "415 UNSUPPORTED_MEDIA_TYPE",
                "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                "417 EXPECTATION_FAILED",
                "418 I_AM_A_TEAPOT",
                "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                "420 METHOD_FAILURE",
                "421 DESTINATION_LOCKED",
                "422 UNPROCESSABLE_ENTITY",
                "423 LOCKED",
                "424 FAILED_DEPENDENCY",
                "425 TOO_EARLY",
                "426 UPGRADE_REQUIRED",
                "428 PRECONDITION_REQUIRED",
                "429 TOO_MANY_REQUESTS",
                "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                "500 INTERNAL_SERVER_ERROR",
                "501 NOT_IMPLEMENTED",
                "502 BAD_GATEWAY",
                "503 SERVICE_UNAVAILABLE",
                "504 GATEWAY_TIMEOUT",
                "505 HTTP_VERSION_NOT_SUPPORTED",
                "506 VARIANT_ALSO_NEGOTIATES",
                "507 INSUFFICIENT_STORAGE",
                "508 LOOP_DETECTED",
                "509 BANDWIDTH_LIMIT_EXCEEDED",
                "510 NOT_EXTENDED",
                "511 NETWORK_AUTHENTICATION_REQUIRED"
              ]
            },
            "message": {
              "type": "string"
            },
            "suppressed": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "stackTrace": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "classLoaderName": {
                          "type": "string"
                        },
                        "moduleName": {
                          "type": "string"
                        },
                        "moduleVersion": {
                          "type": "string"
                        },
                        "methodName": {
                          "type": "string"
                        },
                        "fileName": {
                          "type": "string"
                        },
                        "lineNumber": {
                          "type": "integer",
                          "format": "int32"
                        },
                        "nativeMethod": {
                          "type": "boolean"
                        },
                        "className": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "message": {
                    "type": "string"
                  },
                  "localizedMessage": {
                    "type": "string"
                  }
                }
              }
            },
            "localizedMessage": {
              "type": "string"
            }
          }
        },
        "ForbiddenActionException": {
          "type": "object",
          "properties": {
            "cause": {
              "type": "object",
              "properties": {
                "stackTrace": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "classLoaderName": {
                        "type": "string"
                      },
                      "moduleName": {
                        "type": "string"
                      },
                      "moduleVersion": {
                        "type": "string"
                      },
                      "methodName": {
                        "type": "string"
                      },
                      "fileName": {
                        "type": "string"
                      },
                      "lineNumber": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "nativeMethod": {
                        "type": "boolean"
                      },
                      "className": {
                        "type": "string"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "suppressed": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "stackTrace": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "classLoaderName": {
                              "type": "string"
                            },
                            "moduleName": {
                              "type": "string"
                            },
                            "moduleVersion": {
                              "type": "string"
                            },
                            "methodName": {
                              "type": "string"
                            },
                            "fileName": {
                              "type": "string"
                            },
                            "lineNumber": {
                              "type": "integer",
                              "format": "int32"
                            },
                            "nativeMethod": {
                              "type": "boolean"
                            },
                            "className": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "message": {
                        "type": "string"
                      },
                      "localizedMessage": {
                        "type": "string"
                      }
                    }
                  }
                },
                "localizedMessage": {
                  "type": "string"
                }
              }
            },
            "stackTrace": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "classLoaderName": {
                    "type": "string"
                  },
                  "moduleName": {
                    "type": "string"
                  },
                  "moduleVersion": {
                    "type": "string"
                  },
                  "methodName": {
                    "type": "string"
                  },
                  "fileName": {
                    "type": "string"
                  },
                  "lineNumber": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "nativeMethod": {
                    "type": "boolean"
                  },
                  "className": {
                    "type": "string"
                  }
                }
              }
            },
            "status": {
              "type": "string",
              "enum": [
                "100 CONTINUE",
                "101 SWITCHING_PROTOCOLS",
                "102 PROCESSING",
                "103 EARLY_HINTS",
                "103 CHECKPOINT",
                "200 OK",
                "201 CREATED",
                "202 ACCEPTED",
                "203 NON_AUTHORITATIVE_INFORMATION",
                "204 NO_CONTENT",
                "205 RESET_CONTENT",
                "206 PARTIAL_CONTENT",
                "207 MULTI_STATUS",
                "208 ALREADY_REPORTED",
                "226 IM_USED",
                "300 MULTIPLE_CHOICES",
                "301 MOVED_PERMANENTLY",
                "302 FOUND",
                "302 MOVED_TEMPORARILY",
                "303 SEE_OTHER",
                "304 NOT_MODIFIED",
                "305 USE_PROXY",
                "307 TEMPORARY_REDIRECT",
                "308 PERMANENT_REDIRECT",
                "400 BAD_REQUEST",
                "401 UNAUTHORIZED",
                "402 PAYMENT_REQUIRED",
                "403 FORBIDDEN",
                "404 NOT_FOUND",
                "405 METHOD_NOT_ALLOWED",
                "406 NOT_ACCEPTABLE",
                "407 PROXY_AUTHENTICATION_REQUIRED",
                "408 REQUEST_TIMEOUT",
                "409 CONFLICT",
                "410 GONE",
                "411 LENGTH_REQUIRED",
                "412 PRECONDITION_FAILED",
                "413 PAYLOAD_TOO_LARGE",
                "413 REQUEST_ENTITY_TOO_LARGE",
                "414 URI_TOO_LONG",
                "414 REQUEST_URI_TOO_LONG",
                "415 UNSUPPORTED_MEDIA_TYPE",
                "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                "417 EXPECTATION_FAILED",
                "418 I_AM_A_TEAPOT",
                "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                "420 METHOD_FAILURE",
                "421 DESTINATION_LOCKED",
                "422 UNPROCESSABLE_ENTITY",
                "423 LOCKED",
                "424 FAILED_DEPENDENCY",
                "425 TOO_EARLY",
                "426 UPGRADE_REQUIRED",
                "428 PRECONDITION_REQUIRED",
                "429 TOO_MANY_REQUESTS",
                "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                "500 INTERNAL_SERVER_ERROR",
                "501 NOT_IMPLEMENTED",
                "502 BAD_GATEWAY",
                "503 SERVICE_UNAVAILABLE",
                "504 GATEWAY_TIMEOUT",
                "505 HTTP_VERSION_NOT_SUPPORTED",
                "506 VARIANT_ALSO_NEGOTIATES",
                "507 INSUFFICIENT_STORAGE",
                "508 LOOP_DETECTED",
                "509 BANDWIDTH_LIMIT_EXCEEDED",
                "510 NOT_EXTENDED",
                "511 NETWORK_AUTHENTICATION_REQUIRED"
              ]
            },
            "message": {
              "type": "string"
            },
            "suppressed": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "stackTrace": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "classLoaderName": {
                          "type": "string"
                        },
                        "moduleName": {
                          "type": "string"
                        },
                        "moduleVersion": {
                          "type": "string"
                        },
                        "methodName": {
                          "type": "string"
                        },
                        "fileName": {
                          "type": "string"
                        },
                        "lineNumber": {
                          "type": "integer",
                          "format": "int32"
                        },
                        "nativeMethod": {
                          "type": "boolean"
                        },
                        "className": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "message": {
                    "type": "string"
                  },
                  "localizedMessage": {
                    "type": "string"
                  }
                }
              }
            },
            "localizedMessage": {
              "type": "string"
            }
          }
        },
...
        "CardDTO": {
          "type": "object",
          "properties": {
            "cardId": {
              "type": "integer",
              "format": "int32"
            },
            "type": {
              "type": "string"
            },
            "number": {
              "type": "string"
            },
            "date": {
              "type": "string",
              "format": "date-time"
            },
            "cvv": {
              "type": "integer",
              "format": "int32"
            },
            "isDefault": {
              "type": "boolean"
            }
          }
        },
...
        "IntIdRequestDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
...
        "CardNewDTO": {
          "type": "object",
          "properties": {
            "number": {
              "type": "string"
            },
            "date": {
              "type": "string",
              "format": "date-time"
            },
            "cvv": {
              "type": "integer",
              "format": "int32"
            },
            "isDefault": {
              "type": "boolean"
            }
          }
        },
...
      }
    }
  }